Registered the new component in workbench.common.main.ts and renamed scopeTree.contribution.ts to explorerViewlet.ts
